Improve scp path escape regex.
The previous one reportedly ended up escaping 't' to '\t' in some
setups (at least bash 3.2.48 on OS X 10.6, but not with 4.1.7 on
Fedora nor 3.2.25 on CentOS), probably due to missing backslash in the
regex.  This one removes some spurious escaping and duplicates in the
char class, and should communicate the intent better.