Get rid of bash_completion self-parsing from _filedir_xspec (RedHat: #479936).
My tests show a ~5% (cached) load time improvement with this, it'll
make _filedir_xspec somewhat faster too, and it's generally less ugly.
I suppose we'll rewrite this stuff again in the near future when we
start to take advantage of dynamic completion loading, but until then
I'll take the small improvement.