make: Offer hidden targets when it is clear that the user is trying to complete one of them
Please apply the patch below for "make" completion.

It adds support for completing hidden targets when the user has partially
completed it.

My previous trouble with the unit test is resolved - it is due to my
support for display-mode completion which requires an extra option in the
test case assertion:

   -expect-cmd-minus "."