Use command built-in to run sed to avoid any custom aliases
The nature of bash completions means they're sourced into a user's
existing shell. This means they have to play nicely with any
customisations the user makes. The biggest source of problems are
undoubtedly when a user has a custom alias for a core command.

If you alias 'sed' to 'sed -r' to enable modern regex syntax by default,
it can sometimes cause conflicts. This is the case with the
"_known_hosts_real" helper function provided by the bash_completion
project to all completion functions. Fortunately, this is easy to work
around: just use the command bash built-in to run sed. This avoids the
conflict altogether.