Don't complete hostnames found after Hostname in ~/.ssh/config
When connecting to a host listed after Hostname the options in the
config file are not applied, so don't complete these names accordingly.

If you still want some hosts completed, just add an other entry with
just Host <yourhostname> at the end of the config file.

This closes scop/bash-completion#3 and https://bugs.debian.org/689585