Expand variables to see if directories are handled
that is expand variables and command in backticks to get a usable
command line.  For the ls command and it various aliases allow to
handle files and directories as well.   For this avoid escaped
variables at completion time by switching complete options on demand.
To allow wildcards for the ls command and it aliases the bash
defaults are enabled as fall back.

Side effect is that if variables for directories are used, those
remain expanded on the command line.

This fixes openSUSE bugs #889319, #905348, and #940837
as well as SUSE Linux Enterprise bug #959299.

Signed-off-by: Werner Fink <werner@suse.de>