tipc: handle complete words without trailing space
Prior to this commit,
$ tipc socke<TAB>
$ tipc socket <TAB>
Both produced "tipc socket " while

While
$ tipc socket<TAB>
didn't produce anything. This was true for almost all commands and was
a result of the recipe design. The mechanism that hinder repetition
didn't handle complete words without a trailing space.

In this patch we fix this, so that pressing tab after a completed
valid subcommand inserts a space.

Fixes #60