Improve support of make autocompletion for generated files as produced by Automake.
If the files get too large (> 50000 lines of code in my example),
the calls to make and sed become very slow (make takes about 4 seconds
while sed takes up to 11 seconds, again in my example).

Additionally, there is a lot of noise in the recommendations.
I get about 2800 results, mostly uninteresting.

As a hack, I added a simple cache for expensive calls to make and sed, respectively.
I also filter away the search results that only the binaries are left.