man: use "man -k" (apropos) to find completions
Fixes #73.

The old code attempts to guess manual names from filenames, but this has
a few drawbacks. On one hand the logic to determine the manpath was
incorrect, as explained in #73. On the other hand, not all manual names
correspond to filenames. Both issues are subtle and non-portable as the
exact semantics vary from one man program to another.

The new code uses "man -k" to find potential completions. This method
seems portable and it should be correct -- as long as the man database
is up to date.

The one delicate detail is parsing the output of "man -k". The proposed
version works ok at least with both man-db and mdocml packages.

The issue with manual names not corresponding to filenames is still
broken with man-db, but it seems their bug. It is fixed with mdocml.