(test suite): Fix alias and cd cursor position tests
The intent is to complete a few cursor strokes back into the command
line. But the backward cursor movement cannot be matched by a \b in
expect regexp; it means backspace /delete character on left), not
backward char (move cursor left). Not sure if the expected cursor
position in the completion result can be matched using any regexp. Just
removing the attempt and just leaving the completed string there as is
makes the tests pass, and the completions appear to be doing the right
thing as well.