(test suite): Do not use -ltrim-colon-completions for ssh
This option causes a FAIL in the case when $hosts has only one element.

It may be a bug in assert_complete_one option -ltrim-colon-completions,
but since that option is only used in ssh.exp and the tests seem to pass
without using the option, I don't see the point.