kldunload: Show modules with digits
There were some issues with the sed implementation of kldunload
completions. This is a rewrite, which uses compgen.

It fix the issue with not displaying kernel modules with digits inside
their names, which means that previously a loaded kernel module called
"i915kms" would never be suggested to the user by Bash completions.

Also, thanks to the replacement of sed with compgen, the user will no
longer see errors like "RE error: parentheses not balanced" when
requesting completion after typing "kldunload /".

Closes #184