pkg-config: datadir relative variables
Although `datarootdir` is defined as `${prefix}/share`, the user has
the possibility to define `datadir` in a different directory.

Due to this `completionsdir` and `helpersdir` that are defined in
the generated pkg-config `.pc` file should be relative to `datadir`
variable.

This helps any pkg-config user to set these variables to proper
directories by using the `define-variable` command line argument.