chromium-browser: add more options
Ubuntu and Debian have a shell script that parses a few options from the
manual page. These options are quite limited though and is reflected in
debian/README.debian:

    Chromium has an extensive set of command-line flags that change rapidly, so
    it's not really practical to include documentation in the package.  This page
    is kept relatively up to date with that information:

    http://peter.sh/experiments/chromium-command-line-switches

Arch Linux does not bother with a shell script and provides the binary
where the --help output only mentions a single --password-store option.
Try to be more helpful and extend the current suggested options (from
the manual page) with (extension) developer-oriented options that are
present in my private completion script.