test: fix misinterpretation of completion output in tests
A race condition exists between the program writing data and the moment
where 'expect' reads the output. Travis showed sporadic failures that
indicate reading a partial MAGIC_MARK which was subsequently interpreted
as part of the (possibly empty) completion output.

To avoid this wrong match, always check for the magic marker. Tested
with test_look.py and various versions of /usr/share/dict/words (one
empty file, one with 'foo' and one with 'foo' and 'food' as contents).