Add warning about old bash versions
I spent a good bit of time trying to figure out why I had issues with
bash-completion on macOS 10.12.16, and I finally determined that it was not
loading since the macOS "stock" version of bash is 3.2.57. This will display a
warning to give users a clue that things are not loading successfully.