mutt: Expand sourced filenames, which may contain quotes
The main configuration file for mutt, ~/.muttrc, may source other files.
For instance, some users save email address aliases in `~/.mutt/aliases'
or `~/.aliases'.

The commit

commit 1ffb918b4e39fbc3e89b04822ab58dc39b5656e1
Author: Ville Skyttä <ville.skytta@iki.fi>
Date:   Thu Apr 14 22:17:09 2011 +0300

    Support tildes when recursively sourcing muttrc files (Debian: #615134).

fixed a bug [1] where the tilde character was not correctly expanded,
however, the fix does not work when the filename passed to the source
command is quoted [2].

Ville Skyttä suggested that using `eval' could solve the problem [3].
This patch implements this suggestion.

Testing it can be accomplished with the following instructions:

  $ cat ~/.mutt/aliases1
  alias unquoted \"Someonte Unquoted\" <no@quotes.org>
  $ cat ~/.mutt/aliases2
  alias quoted \"Someone Quoted\" <yes@quotes.org>

  $ cat ~/.muttrc
  source ~/.mutt/aliases1
  source "~/.mutt/aliases2"

Before this patch, the following completion attempt did not work:

  $ mutt -A [TAB]
  $ mutt -A unquoted [CURSOR]
  (wrong, should have suggested both aliases)

After this patch, it works:

  $ mutt -A [TAB]
  quoted    unquoted

[1] https://bugs.debian.org/615134
[2] https://bugs.debian.org/615134#25
[3] https://bugs.debian.org/615134#30