Reset state of hostcomplete shopt after local unset
When hostcomplete option is set, bash will attempt to perform hostname
completion for words starting with `@', as describe in the manual page
for bash (1):

    hostcomplete
            If set, and readline is being used, bash will attempt to
            perform hostname completion when a word containing  a  @
            is   being  completed  (see  Completing  under  READLINE
            above).  This is enabled by default.

For instance:

  $ @[TAB]
  @::1   @localhost

However, when bash-completion is installed, such completions do not
work, and that's because bash_completion unsets hostcomplete before
installing completions for finger, ssh, sftp, among others, but never
restores the original setting.  This patch adds a save/restore step.

This bug has been reported for Debian [1].

[1] https://bugs.debian.org/525082