test: avoid interrupting magic mark output
The same race condition as 85275cf5 exists for multiple completions
(case 0).  If the magic mark is not included in `expect()`, it may or
may not have been written by bash at the point `bash.sendintr()` is
called.   If the interrupt is delivered before magic mark has finished
printing, it can corrupt the prompt and cause `bash.expect_exact(PS1)`
to timeout.

Include `MAGIC_MARK` in the case 0 regex to avoid this issue.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>