test: Increase expect pty to 160 columns
Bash wraps lines based on the size of the pseudo-terminal used by
expect.  If a completed command + magic mark is longer than 80-${#PS1}
characters it will insert "\r\n" (and usually several spaces and
backspaces) at the wrap point which can corrupt the magic mark.

Ideally, wrapping would either be avoided in all cases or handled
robustly by expect in all cases, but since this is no easy task,
increase the number of columns to reduce wrapping in common cases.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>