test for _xspecs declaration
Test if _xspecs is declared before to use xspec completion to avoid syntax error (invalid arithmetic operand) when the command contains non word character ([^a-zA-Z0-9_]). Typically, $cmd with a dot or an hyphen or a non ascii char triggers this syntax error when completing args because if _xspecs is not explicitely declared, ${_xspecs[$cmd]} is interpreted as a simple rather than associative array, and so $cmd in this context has to be a valid variable name rather than a keyword.