perl: fix completion with space between option and argument
Most perl options do not allow a space between the switch and the
argument, however, bash-completion incorrectly suggests completions for
`-d', `-M` (or `-m'), and `-V', as reported in Debian bug #614775 [1],
as well as for `-x', which was not mentioned in the bug report.  This
patch fixes the incorrect completions for these options, by falling back
to regular _filedir completion when a space is present.

On the other hand, and unlike all other perl options, a space between
the `-e', `-E', and `-I' options and their arguments, e.g. `perl -e
"exit 2"', *is* valid syntax.  However, the argument for `-e' and `-E'
is neither a filename nor a directory, but one line of perl program.
So, in order to keep the old behavior, which was correct, this patch
adds extra code that skips _filedir completion for `-e' and `-E' with a
space.

Finally, the test case for `-x' requires changes, because it expects
that `-x' with a space should complete only with directories.  That is
misleading, because it might suggest that the directory would be treated
as the argument to the `-x' option, when it will not, as can be verified
with the following test:

  $ cat test.pl
  Some regular text with an embedded script
  #!perl
  use Cwd;
  my $dir = getcwd;
  print "$dir\n";
  __END__
  More text messages, not code.

By executing the script with the `-x[dir]' option, the non-code part of
the file will be ignored, i.e. only the embedded script will be
executed.  It will also be executed under [dir], if it is provided:

  $ perl -xperl/ test.pl
  /tmp/perl

However, if a space is present between `-x' and [dir], the script fails,
because the perl interpreter thinks that [dir] is a script, not the
argument to `-x':

  $ perl -x perl/ test.pl
  Can't open perl script "perl/": Is a directory

The rationale for the changes is based on `perl --help', which explains
that there are no spaces between options and arguments (see note below
for `-I'):

  $ perl --help | grep "\-d\|\-V\|-I\|-x\|-\[mM\]"
    -d[:debugger]     run program under debugger
    -Idirectory       specify @INC/#include directory (several -I's allowed)
    -[mM][-]module    execute "use/no module..." before executing program
    -V[:variable]     print configuration summary (or a single Config.pm variable)
    -x[directory]     ignore text before #!perl line (optionally cd to directory)

When a space is required, `perl --help' makes it explicit:

  $ perl --help | grep "\-e\|-E"
    -e program        one line of program (several -e's allowed, omit programfile)
    -E program        like -e, but enables all optional features

As an exception to this rule, `-I' does accept a space. ¬¬

[1] https://bugs.debian.org/614775