_init_completion: set default $IFS
Many completions expect IFS to be the default value. For example,
something like

    compgen -W 'foo bar' -- "$cur"

or

    opts='foo bar'
    for i in $opts; do ...

will not work as expected if IFS is set in the shell to a value that
does not contain space.

However, functions don't seem set IFS if they are written to work with
the default value. Setting it to the default in `_init_completion` should
help ensure they run as expected. Since the variable is `local`, it will
have no effect on the outside shell.