apt-cache: avoid nonzero exit code from _apt_cache_packages
On systems where apt-cache is not available but we attempt an apt-get
command that does "_xfunc apt-cache _apt_cache_packages" (no matter if
apt-get is installed) to add to some populated completions in COMPREPLY,
for some reason the failure empties whatever is in COMPREPLY even though
we're using +=. Current apt-get build-dep completion is an example of
this on e.g. CentOS 6 which has neither apt-get or apt-cache, dirs
should be completed even though apt-cache is not there. Haven't
investigated why, but this works around it.