feat(jq): support all command-line options
The options are not all listed by --help, so look in the man page to
find a nearly complete list (I hope!). Neither the man page nor --help
mentions --help or --version; I have not looked elsewhere to see if
there might be yet more.

Guard against the possibility of jq being fixed in future by testing to
see if its --help output contains “--help” and using it if so,
otherwise, use the hard-coded list; see
https://github.com/stedolan/jq/issues/2284