test(diff_env): detect unwanted changes of OLDPWD in test
To properly detect unwanted changes of OLDPWD by completions, we
remove OLDPWD from the white list of mutable variables.  Now, we need
to properly save and restore the value of OLDPWD when we change the
testing directory on purpose.

We add variable names "_bash_completion_test_*" in the white list and
use "_bash_completion_test_OLDPWD" to save the value of "OLDPWD".

Remark: Another suggested solution was to use "pushd" and "popd" in
order to save/restore the directory, but this idea was discarded
because they turned out to affect "OLDPWD" too.