test(bash_env_saved): correctly restore "cwd"
In the previous version of `bash_env_saved', `bash.cwd' was used to
retrieve the current working directory of the Bash process, but it
turned out that `bash.cwd' just retains the value of the startup time
of the process but not the current one.  Instead we shall store the
real current working directory in a shell variable inside the Bash
process.