test(bash_env_saved): work around nested bash_env
In the previous version of `bash_env_saved', when the same variables
are saved or restored in the nested `with bash_env_saved()' statement,
the variables are not correctly restored.  It is not clear whether
there are real instances of such cases in existing tests, but
`bash_env_saved' shall support the nested call of `with
bash_env_saved()' to avoid future troubles.