fix(_ltrim_colon_completions): do compaction of sparse COMPREPLY
The existing implementation of "_ltrim_colon_completions()" assumes
that COMREPLY is a dense array, i.e., all the array subscripts are
within the range 0 .. ${#COMPREPLY[@]}-1.  However, the highest
assigned subcript i can be i >= ${#COMPREPLY[@]} when the array is
sparse.  In fact, COMPREPLY can be sparse in "_known_hosts_real()"
because it unsets an empty elements before calling
"_ltrim_colon_completions()".