fix(secret-tool): fix empty candidates for "secret-tool search" options
In the completion of `secret-tool', the option names `--all' and
`--unlock' are stored as the keys of the associative array `opts' but
not as the values.  In addition, the test had been wrongly requested
not to produce any completion results for "secret-tool search ".  Now
we have the proper expected result of the test.