fix(secret-tool): protect unset opts[...] from pathname expansions
``unset opts["$word"]'' will be unintendedly expanded to `opts<X>'
when $word contains a character <X> and there is a file named
`opts<X>'.  Even if there is no matching filenames, when `shopt -s
failglob' is set, ``unset opts["$word"]'' always fails.  We should
quote it as "unset 'opts[$word]'".  In addition, it is better to
specify `-v' option in case that there is a confusing name of a shell
function.