feat(pylint): message ids completion
Complete message ids for `--fail-on`, `--help-msg`, `--enable`, and
`--disable`. We scrape the output of `--list-msgs-enabled` instead of
`--list-msgs`, because it outputs both enabled and disable message ids
and thus allows for one implementation for all cases, even though the
output is somewhat less parseable than that of the latter.

Curiously, `--list-msgs-enabled` also outputs some message ids that
`--list-msgs` does not.