ci(renovate): use major `regex` versioning with perltidy
The default `semver` isn't applicable to perltidy's tags, which are
`yyyymmdd(.nn)`. `loose` would capture all those, but we'd get updates
for both the "major" i.e. `yyyymmdd` stream, as well as for the "micro"
i.e. `.nn` versions. That could be changed with `separateMajorMinor`,
but we're not really interested in the "minor" changes here as the
version component semantics do not match the usual major/minor -- I
believe "minor" here is really a developer post-release test version.

Use a regex to look only at the `yyyymmdd` part.

Caveat: this would break if we'd sometime have a `yyyymmdd.nn` version
in our pre-commit config, but that seems unlikely and this is the best I
can come up with at the moment.