ci(pre-commit): run checks on commit messages
This is a bit involved, but the gist of it is to dump each commit
message to a file in turn, and run commitlint through pre-commit on it.

This could conceivably be done using the CLI git log, but I failed to
come up with a way to produce a nice range of commits that'd work
reliably for pull requests. Pushes are somewhat simpler. But both them
and pull requests require deepening the shallow clone GitHub does by
default, and for pull requests determining exactly by how much would
have required some scriptwork on the actions contexts and GitHub REST
API anyway. So while we're there, do it all using the API's/contexts and
do not add git CLI to the mix at all.