ci: run release-please only on pushes
Running it on pull requests is not needed when the source repo is the
main one, because we only want the change log updated when changes land
in the default branch.

On the other hand, when the pull request source repo is something else
besides the main one, trying to run it causes permission denials and CI
failures.

Run it only on pushes to resolve both issues.

Note: assumes that we limit the branches on which we run on pushes, in
practice only to the default one.