style: explicitly specify the type of "unset" by the option "-v" or "-f"
If "unset" is called without the target type "-v" (variable) or "-f"
(function), it may remove unexpected items.  We can always explicitly
specify the type to avoid it.

Note that, even though some form of the identifiers such as 'arr[key]'
is always treated as a variable name when the type is unspecified, we
also specify the type in such cases to easily detect the other needed
cases by regex 'unset [^-]'.