Add completion script for dgit
This script autocompletes the most common subcommands, but only provides
arguments autocompletion for only two of them, import-dsc and rpush, for
practical reasons.

For dgit options, it provides two-level completion for the most popular
options, excluding rarely used or dangerous ("use when you know what
you’re doing") options.  The two-levels work in a way to not overwhelm
the user with options when they Tab twice, so only some but not all
arguments to e.g. --quilt= are shown:

$ dgit --
--baredebian                                 --force-import-gitapply-no-absurd
--baredebian+git                             --force-reusing-version
--baredebian+tarball                         --force-unrepresentable
--build-products-dir=                        --force-unsupported-source-format
--clean                                      --force-uploading-binaries
--clean=dpkg-source                          --force-uploading-source-only
--clean=git                                  --gbp
--clean=git-ff                               --include-dirty
--damp-run                                   --new
--delayed=                                   --no-rm-on-error
--deliberately-fresh-repo                    --no-sign
--deliberately-include-questionable-history  --overwrite
--deliberately-not-fast-forward              --package
--dep14tag                                   --quilt
--distro=                                    --quilt=auto
--dpm                                        --quilt=linear
--dry-run                                    --rm-old-changes
--force-changes-origs-exactly                --save-dgit-view=
--force-dsc-changes-mismatch                 --split-view=
--force-import-dsc-with-dgit-field           --upstream-commitish=
--force-import-gitapply-absurd

However, when the user types --quilt=, the rest of them are shown:

$ dgit --quilt=
auto            baredebian+git      dpm      linear       nofix      unapplied
baredebian      baredebian+tarball  gbp      nocheck      smash

For --upstream-commitish= the script autocompletes some tags, starting
with tags with the matching version, and the rest sorted by version:

$ dgit --upstream-commitish=upstream/1.
upstream/1.11.0  upstream/1.6.0

For options accepting both = and space (--quilt, --clean, --package)
the completion is performed for both versions; the list of all options
above shows the "spaced" (--clean) version and a couple of commonly
used versions with = and an argument. For options only accepting =,
the option ends with = to allow the user to start typing the argument
directly after that.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>