fix(man,info): remove ineffective "_expand" calls
The function "_expand" processes "cur" that begins with a tilde, which
is actually already processed by earlier codes in these completions.
Thus "_expand" will never see "cur" starting with a tilde there.

The earlier codes processing the tilde were introduced in commit
1b85a1b.  The calls of "_expand" was moved after the code in commit
add4e3c, for which we forgot the background unfortunately.  We here
just remove the calls of "_expand".