fix: Properly handle environments with unset IFS
Prior to this commit, the current state of `IFS` was stored into a local
variable `ifs`. Later, IFS is restored by setting `IFS=$ifs`. If the
user starts with either a set `IFS` or an empty `IFS`, then this
behavior is correct. If the user starts with an unset `IFS`, then
`IFS=$ifs` results in an empty `IFS`, which is not the original state.
Most of the time an empty `IFS` is not desired, and indeed it breaks
some things inside bash-completions. eg

```
unset IFS
local ifs=$IFS IFS=$' \t\n'
restore=$(shopt -po noglob)
IFS=$ifs
$restore
...
-bash: set -o noglob: command not found
```

With this commit, all instances of `local ifs=$IFS` have been removed.
Instead, the `local IFS` is dynamically unset.