refactor(_umount.linux): remove unnecessary IFS setups
This removes some ineffective IFS settings IFS=$'\0' found in #687
[1].  The quoted strings are not subject to the word splitting, where
IFS does not have effects.  The setting IFS=$'\0' has been first
introduced in 99b28947, and the related codes have been updated in
1c4b4618, 2f61acd0, 4375c4b9, and f2412520.  In the change history,
the setting IFS=$'\0' does not seem to have had any effect at any
point, so we just remove the setup IFS=$'\0' today.

The original intent is actually not so clear.  The related code has
been first introduced to process octal escape sequences of the form
`\ooo` (where `o` is a digit 0-7) used in `/etc/fstab` [2].  One
possibile intent might have been to split the filesystem names by NUL
represented by `\000` in `/etc/fstab`, but we cannot find any
information on the special meaning of NUL in `/etc/fstab`.  We think
the intent has probably been to make sure that the word splitting
would not happen, which was actually not needed for the quoted
strings.

[1] https://github.com/scop/bash-completion/pull/687
[2] man fstab (found online at e.g. https://linux.die.net/man/5/fstab)