_known_hosts_real: added host-exclusion-patterns and docs
This adds host-exclusion-patterns, a functionality that allows users to specify
patterns for which matching hosts shall be excluded from completion.

The use case if one has many known hosts, which cannot be dropped from their
sources (for example because they need special SSH configuration), but which
disturb in day-to-day completion.
Examples:
- A large list of hosts `www.*`, which are rarely used, but “hide” any other
  hosts starting with `w` or make them at least more cumbersome to complete.
- Servers, reachable via `<hostname>.srv.example.org` (which are often needed
  for completion), and their respectively corresponding service processors,
  reachable via `<hostname>.sp.example.org` (which are rarely needed for
  completion).
  Completing `<hostname>` then wouldn't complete the full name, but stop at
  `<hostname>.s`.

This functionality allows to exclude such hosts from completion, and yet
provides a easy way to temporarily switch them on again (for example by creating
a small shell script, that enables/disables the respective pattern).
When the pattern are read from file (and not from a shell variable), this
switching can then be easily made working even across shells.

Implementation Details:
- The code should work with `set -e` and/or `set -u`. Current versions of bash
  seem to expand "${BASH_COMPLETION_HOST_EXCLUSION_PATTERN_FILES[@]}" even if
  not set, but for the case this ever changes, the used
  "${BASH_COMPLETION_HOST_EXCLUSION_PATTERN_FILES[@]-}" seems to work as
  intended, too.
  AWK is specified to ignore empty arguments, so this should be a problem there.
- The could should hopefully be independent of `IFS`.
- AWK (`awk 1 - "${BASH_COMPLETION_HOST_EXCLUSION_PATTERN_FILES[@]-}"`) is used
  to add any possibly missing newline terminators to the final line in each file
  respectively the value of the variabl `BASH_COMPLETION_HOST_EXCLUSION_PATTERN`
  read from standard input.

Code was tested with GNU grep as well as GNU awk, MAWK and Original-AWK.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>