feat(7z): complete 7zr and 7zz
p7zip provides [7zr(1)] in addition to [7za(1)] and [7z(1)].  7zr
behaves like 7z with the exception that it only supports the 7z, LZMA,
and XZ formats.

Since [7-Zip 21.01 alpha], the 7-Zip project provides a command-line
version of 7-Zip for Linux, [7zz(1)].  It shares nearly all of the
command-line options and syntax of the p7zip port.

Apply the current completions for 7z to 7zr and 7zz.  Note that there
are a few minor differences (e.g. 7zz does not support `-mh*` or
`-ms*`) which will need to be addressed by subsequent work.

[7-Zip 21.01 alpha]: https://sourceforge.net/p/sevenzip/discussion/45797/thread/d401ab2966/
[7z(1)]: https://github.com/jinfeihan57/p7zip/blob/v17.04/man1/7z.1
[7za(1)]: https://github.com/jinfeihan57/p7zip/blob/v17.04/man1/7za.1
[7zr(1)]: https://github.com/jinfeihan57/p7zip/blob/v17.04/man1/7zr.1
[7zz(1)]: https://manpages.debian.org/testing/7zip/7zz.1.en.html

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>