fix(gcc): explicitly check existence of the commands
When at least one of the commands "cc", "c++", "f77", and "f95" is not
found in the system, on the first load of "gcc" completions,
"command_not_found_handle" hook (typically set by distributions) is
invoked in the middle of completion, breaks the terminal layout by
outputting messages, and eats user inputs.  This commit fixes the
issue by checking the existence of commands before running these
commands.

Refs.
https://github.com/scop/bash-completion/pull/390
https://github.com/akinomyoga/ble.sh/issues/192
https://github.com/akinomyoga/ble.sh/issues/203