fix(make): deactivate the display-only mode
In the display-only mode where COMP_TYPE=9, 37, or 42, only a part of
the completion is stored in COMPREPLY for displaying purposes.  For
example, "xyz" is stored in COMPREPLY when "abc/xyz" is the candidate
and "abc/" is already inserted.

However, the test framework extracts generated completions using
COMP_TYPE=37 by setting "set show-all-if-ambiguous off", which would
be broken by the display-only mode.  There is no simple way to make it
work with the test framework, and the display-only mode does not seem
to be essential.  For the time being, we deactivate the display-only
mode.  See also discussions in the following links:

https://github.com/scop/bash-completion/issues/544
https://github.com/scop/bash-completion/pull/546