fix(brctl): switch handling of show and addbr
The show subcommand accepts a bridge name as a paramenter so let's
offer completion.  On the other hand addbr creates new interface and
requires a name that doesn't exist yet.

Longer awk condition prevents offering interfaces that are parts of
bridges. They are all displayed in the last column but with all other
columns being blang on the second and following lines of each bridge
the last column is also the first.