chore(update-{deprecated-links,test-cmd-list},runLint): update shebang
The test scripts have had the shebang #!/bin/bash, but Bash is not
necessarily located at /bin/bash.  For example, it does not located
there in FreeBSD and MINIX.  As we require /bin/bash in the shebang of
the pre-commit scripts, any attempts to commit changes in such a
system would be blocked by the pre-commit hook.

In this patch, we update the shebang to `#!/usr/bin/env bash`, which
is a widely used workaround for this issue.

Note: I have not changed the shebang of
`test/docker/*/install-packages.sh` as these seem to be always called
in a GNU/Linux container, where /bin/bash exists.