fix(strace): fix the test for an associative array
The original test seems to try to check whether the array `syscalls`
is non-empty, but this does not work for associative arrays.  The
expansion ${array_name} references ${array_name[0]}, but
${array_name[0]} does not necessarily exist even when the array is
non-empty when the array is an associative array or a sparse array.

We can explicitly check the number of elements.