fix(__load_completion): don't resolve symlinks to invoked command
This might make the path we install completions for different from the
invoked path so that the completion would fail to actually load after
setup.

The final path segment might also have a different name in the resolved
symlink, and we should not install it for the resolved name, because the
target's functionality might vary based on the invocation name.

Simply replacing the final path segment with the original basename could
yield a path that does not exist, even though the original invoked one
does.

https://github.com/scop/bash-completion/pull/924#discussion_r1163249029