fix(gnokii): avoid non-POSIX ERE \b
The word-boundary anchor \b is not supported by POSIX Extended Regular
Expressions (ERE).  Also, the previous implementation did not consider
the word-boyndary for the last element.  We here explicitly match the
trailing /[^_[:alnum:]]/ or /$/.

Another issue is that the previous implementation did not work because
of contaminated newlines in the ERE.  We now split the results by
newlines and then join the elements with IFS='|'.